+++ app/components/dialogs/DialogFrame/index.scss
@@ -1,14 +1,13 @@
-
 .Dialog {
   $border-radius: 8px;
 
   position: relative;
-  min-width: 340px;
+  min-width: 300px;
   max-width: 100%;
   color: #333;
   background: #fff;
   border-radius: $border-radius;
-  box-shadow: 0 0 19px 3px rgba(0,0,0, 0.2);
+  box-shadow: 0 0 19px 3px rgba(0, 0, 0, 0.2);
 
   &__close {
     position: absolute;
@@ -58,7 +57,7 @@
     display: flex;
     height: 50px;
     border-radius: 0 0 $border-radius $border-radius;
-    box-shadow: 0 -2px 14px 0 rgba(0,0,0, 0.07);
+    box-shadow: 0 -2px 14px 0 rgba(0, 0, 0, 0.07);
     overflow: hidden;
   }
 }
