+++ app/components/pages/SubmitPost.scss
@@ -11,6 +11,10 @@ body.submit-page {
   }
 }
 
+.App {
+  overflow: hidden;
+}
+
 .SubmitPost {
   padding: 25px 0 1rem;
 }
