+++ src/app/components/header/AccountInfoMobile/index.js
@@ -1,0 +1,1 @@
+export { default } from './AccountInfoMobile';
