+++ src/app/containers/login/LoginForm/LoginForm.jsx
@@ -32,7 +32,7 @@ const Wrapper = styled.div`
     @media (max-width: 500px) {
         width: 100%;
         margin: 10px;
-        min-width: 320px;
+        min-width: 300px;
     }
 `;
